Move pkg/kubectl/proxy to staging
