+++ pkg/kubectl/cmd/proxy/BUILD
@@ -6,9 +6,9 @@ go_library(
     importpath = "k8s.io/kubernetes/pkg/kubectl/cmd/proxy",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/kubectl/proxy:go_default_library",
         "//staging/src/k8s.io/cli-runtime/pkg/genericclioptions:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/cmd/util:go_default_library",
+        "//staging/src/k8s.io/kubectl/pkg/proxy:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/util/i18n:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/util/templates:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
