+++ staging/src/k8s.io/kubectl/pkg/proxy/BUILD
@@ -1,9 +1,19 @@
-package(default_visibility = ["//visibility:public"])
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-    "go_test",
+go_library(
+    name = "go_default_library",
+    srcs = ["proxy_server.go"],
+    importmap = "k8s.io/kubernetes/vendor/k8s.io/kubectl/pkg/proxy",
+    importpath = "k8s.io/kubectl/pkg/proxy",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//staging/src/k8s.io/apimachinery/pkg/util/net:go_default_library",
+        "//staging/src/k8s.io/apimachinery/pkg/util/proxy:go_default_library",
+        "//staging/src/k8s.io/client-go/rest:go_default_library",
+        "//staging/src/k8s.io/client-go/transport:go_default_library",
+        "//staging/src/k8s.io/kubectl/pkg/util:go_default_library",
+        "//vendor/k8s.io/klog:go_default_library",
+    ],
 )
 
 go_test(
@@ -13,20 +23,6 @@ go_test(
     deps = [
         "//staging/src/k8s.io/apimachinery/pkg/util/proxy:go_default_library",
         "//staging/src/k8s.io/client-go/rest:go_default_library",
-    ],
-)
-
-go_library(
-    name = "go_default_library",
-    srcs = ["proxy_server.go"],
-    importpath = "k8s.io/kubernetes/pkg/kubectl/proxy",
-    deps = [
-        "//staging/src/k8s.io/apimachinery/pkg/util/net:go_default_library",
-        "//staging/src/k8s.io/apimachinery/pkg/util/proxy:go_default_library",
-        "//staging/src/k8s.io/client-go/rest:go_default_library",
-        "//staging/src/k8s.io/client-go/transport:go_default_library",
-        "//staging/src/k8s.io/kubectl/pkg/util:go_default_library",
-        "//vendor/k8s.io/klog:go_default_library",
     ],
 )
 
@@ -41,4 +37,5 @@ filegroup(
     name = "all-srcs",
     srcs = [":package-srcs"],
     tags = ["automanaged"],
+    visibility = ["//visibility:public"],
 )
